+++ cmd/start/main_test.go
@@ -189,33 +189,6 @@ func TestApplyVersionTags(t *testing.T) {
 	}
 }
 
-func TestPipelineRunURL(t *testing.T) {
-	tests := map[string]struct {
-		consoleURL string
-	}{
-		"base URL without trailing slash": {
-			consoleURL: "https://console.example.com",
-		},
-		"base URL with trailing slash": {
-			consoleURL: "https://console.example.com/",
-		},
-	}
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			ctxt := &pipelinectxt.ODSContext{Namespace: "foo"}
-			opts := options{consoleURL: tc.consoleURL, pipelineRunName: "bar-ab12c"}
-			u, err := pipelineRunURL(ctxt, opts)
-			if err != nil {
-				t.Fatal(err)
-			}
-			want := "https://console.example.com/k8s/ns/foo/tekton.dev~v1beta1~PipelineRun/bar-ab12c/"
-			if u != want {
-				t.Fatalf("want: %s, got: %s", want, u)
-			}
-		})
-	}
-}
-
 func lastTagPayload(t *testing.T, srv *testserver.TestServer) bitbucket.TagCreatePayload {
 	req, err := srv.LastRequest()
 	if err != nil {
